requirements.txt
amclient==1.1.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.30
protobuf==3.20.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
